+++ src/main/java/org/patikadev/orderexample/service/implementation/CouponServiceImpl.java
@@ -2,12 +2,11 @@ package org.patikadev.orderexample.service.implementation;
 
 import lombok.RequiredArgsConstructor;
 import org.patikadev.orderexample.converter.CouponDtoConverter;
-import org.patikadev.orderexample.converter.CustomerDtoConverter;
 import org.patikadev.orderexample.dto.request.CreateCouponDto;
 import org.patikadev.orderexample.dto.response.CouponResponseDto;
+import org.patikadev.orderexample.exception.ServiceOperationException;
 import org.patikadev.orderexample.repository.CouponRepository;
 import org.patikadev.orderexample.service.CouponService;
-import org.patikadev.orderexample.service.CustomerService;
 import org.springframework.stereotype.Service;
 
 @RequiredArgsConstructor
@@ -15,8 +14,6 @@ import org.springframework.stereotype.Service;
 public class CouponServiceImpl implements CouponService {
     private final CouponRepository couponRepository;
     private final CouponDtoConverter couponDtoConverter;
-    private final CustomerService customerService;
-    private final CustomerDtoConverter customerDtoConverter;
 
     @Override
     public void addCoupon(CreateCouponDto createCouponDto) {
@@ -25,11 +22,20 @@ public class CouponServiceImpl implements CouponService {
 
     @Override
     public CouponResponseDto getCouponById(Long couponId) {
-        return couponDtoConverter.convertToResponseDto(couponRepository.getCouponById(couponId));
+        return couponDtoConverter.convertToResponseDto(couponRepository.findById(couponId)
+                .orElseThrow(() -> new ServiceOperationException.ProductNotFoundException("Coupon not found")));
     }
 
     @Override
-    public void deleteCouponById(Long couponId) {
+    public Boolean existsCouponById(Long couponId) {
+        return couponRepository.existsById(couponId);
+    }
+
+    @Override
+    public void deleteCouponsById(Long couponId) {
+        if (!couponRepository.existsById(couponId)) {
+            throw new ServiceOperationException.CouponNotFoundException("coupon not found");
+        }
         couponRepository.deleteById(couponId);
     }
 }
